Refactor test from async to sync
Reason for change: This test has been flaky on our CI due to its async
nature.

The expectation is for .build to construct TextEditor with the right
params.

We don't have to load the correct language package for this test. We can
load a dummy languageMode and assert that the correct grammar is set on
TextEditor